w3c-reports/12092014/mac-wk2/runned-tests/XMLHttpRequest/send-entity-body-get-head-async-expected.txt


PASS XMLHttpRequest: send() - non-empty data argument and GET/HEAD - async, no upload events should fire (GET) 
FAIL XMLHttpRequest: send() - non-empty data argument and GET/HEAD - async, no upload events should fire (HEAD) assert_equals: expected "NO" but got "0"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
assert_equals@http://localhost:8800/resources/testharness.js:335:15
http://localhost:8800/XMLHttpRequest/send-entity-body-get-head-async.htm:26:26
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/XMLHttpRequest/send-entity-body-get-head-async.htm:25:20)
